Use helpful typer and invoke for root cli commands
