+++ src/mcp_agent/cli/cloud/main.py
@@ -7,11 +7,7 @@ from pathlib import Path
 from typing import Optional
 from importlib.metadata import version as metadata_version
 
-import click
 import typer
-from rich.console import Console
-from rich.panel import Panel
-from typer.core import TyperGroup
 
 from mcp_agent.cli.cloud.commands import (
     configure_app,
@@ -40,8 +36,7 @@ from mcp_agent.cli.cloud.commands.servers import (
     describe_server,
     delete_server,
 )
-from mcp_agent.cli.exceptions import CLIError
-from mcp_agent.cli.utils.ux import print_error
+from mcp_agent.cli.utils.typer_utils import HelpfulTyperGroup
 
 # Setup file logging
 LOG_DIR = Path.home() / ".mcp-agent" / "logs"
@@ -63,34 +58,6 @@ file_handler.setFormatter(
 logging.basicConfig(level=logging.INFO, handlers=[file_handler])
 
 
-class HelpfulTyperGroup(TyperGroup):
-    """Typer group that shows help before usage errors for better UX."""
-
-    def resolve_command(self, ctx, args):
-        try:
-            return super().resolve_command(ctx, args)
-        except click.UsageError as e:
-            click.echo(ctx.get_help())
-
-            console = Console(stderr=True)
-            error_panel = Panel(
-                str(e),
-                title="Error",
-                title_align="left",
-                border_style="red",
-                expand=True,
-            )
-            console.print(error_panel)
-            ctx.exit(2)
-
-    def invoke(self, ctx):
-        try:
-            return super().invoke(ctx)
-        except CLIError as e:
-            # Handle CLIError cleanly - show error message and exit
-            logging.error(f"CLI error: {str(e)}")
-            print_error(str(e))
-            ctx.exit(e.exit_code)
 
 
 # Root typer for `mcp-agent` CLI commands
